Improves deduplication algos

+++ src/main/scala/com/vorlov/commands/Normalize.scala
@@ -3,15 +3,26 @@ package com.vorlov.commands
 import java.io.File
 import java.nio.file.StandardOpenOption
 
+import com.vorlov.api.twitter.model.Tweet
 import com.vorlov.classifier.naivebayes.NaiveBayesClassifier
+import com.vorlov.helper.similarity.CircularSimhashDeduplication
 import com.vorlov.util.IOUtils
 import org.slf4j.LoggerFactory
 import com.vorlov.helper.format.csv._
+
 import com.vorlov.util.Csv._
+import com.vorlov.util.TwitterTokenizer._
 
 object Normalize extends Command {
 
-  val log = LoggerFactory.getLogger(getClass.getName)
+  implicit val log = LoggerFactory.getLogger(getClass.getName)
+  
+  def tweetIDDeduplicate(tweets: Stream[Tweet]): Stream[Tweet] = tweets.sortBy(_.id).sliding(2).filter{
+    _ match {
+      case Stream(t1, t2) if t1.id == t2.id => false
+      case _ => true
+    }
+  }.map(_.head).toStream
 
   val classifier = new NaiveBayesClassifier
 
@@ -22,18 +33,13 @@ object Normalize extends Command {
 
   val tweets = new File(inputPath).asCSV.toStream
 
-  val c = tweets.sortBy(_.id).sliding(2).filter{
-    _ match {
-      case Stream(t1, t2) if t1.id == t2.id => false
-      case _ => true
-    }
-  }.map(_.head)
-
-  IOUtils.delete(outputPath)
+  if(IOUtils.exists(outputPath)) IOUtils.delete(outputPath)
 
   log.info(s"Writing data to $outputPath")
 
-  c.asCSV.foreach {
+  implicit def tweetToString(tweet: Tweet): String = tweet.text.tokens.mkString
+
+  CircularSimhashDeduplication.deduplicate(tweetIDDeduplicate(tweets)).toIterator.asCSV.foreach {
     line =>
       IOUtils.write(outputPath, line + "\n", StandardOpenOption.APPEND, StandardOpenOption.CREATE)
   }
